+++ firmware/tsa_sorter/logger.h
@@ -12,16 +12,16 @@ public:
     LogMode mode() const { return _mode; }
 
     void stateChange(const char* from, const char* to);
-    void classified(uint8_t brickNum, BrickCategory cat, uint8_t targetBin, uint8_t samples);
-    void senseVerbose(uint8_t brickNum, const SenseResult& s);
-    void routeReady(uint8_t brickNum, uint8_t targetBin, uint16_t servoAngle, bool ok,
+    void classified(uint16_t brickNum, BrickCategory cat, uint8_t targetBin, uint8_t samples);
+    void senseVerbose(uint16_t brickNum, const SenseResult& s);
+    void routeReady(uint16_t brickNum, uint8_t targetBin, uint16_t servoAngle, bool ok,
                     const char* positionLabel);
-    void handoffDone(uint8_t brickNum, uint8_t targetBin);
-    void binConfirm(uint8_t brickNum, uint8_t expectedBin, uint8_t actualBin,
+    void handoffDone(uint16_t brickNum, uint8_t targetBin);
+    void binConfirm(uint16_t brickNum, uint8_t expectedBin, uint8_t actualBin,
                     uint32_t transitMs, bool ok);
     void thermal();
-    void runComplete(uint32_t totalMs, const uint8_t counts[4], bool countsMatch);
-    void errorHalt(uint8_t brickNum, uint8_t expectedBin, const char* reason);
+    void runComplete(uint32_t totalMs, const uint16_t counts[4], bool countsMatch);
+    void errorHalt(uint16_t brickNum, uint8_t expectedBin, const char* reason);
     void info(const char* msg);
 
 private:
@@ -30,7 +30,7 @@ private:
     void printCsvHeader();
     void printHumanPrefix(const char* label);
     void printCsvRow(const char* eventName,
-                     uint8_t brickNum,
+                     uint16_t brickNum,
                      const char* stateName,
                      const char* category,
                      uint8_t targetBin,
